# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nature Writing Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: articles
  - name: categories
  - name: tags
  - name: species
  - name: portfolio
  - name: contact
  - name: newsletter
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Article"
                  total:
                    type: integer
                required: [articles, total]
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create an article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

  /articles/featured:
    get:
      operationId: getFeaturedArticles
      tags: [articles]
      summary: Get featured articles for the homepage
      responses:
        "200":
          description: Featured articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"

  /articles/stats:
    get:
      operationId: getArticleStats
      tags: [articles]
      summary: Get article statistics grouped by category
      responses:
        "200":
          description: Article stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleStats"

  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get an article by slug (also increments view count)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by slug with its articles
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category with articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithArticles"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagBody"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"

  /tags/{id}:
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /species:
    get:
      operationId: listSpecies
      tags: [species]
      summary: List all species
      parameters:
        - name: kingdom
          in: query
          schema:
            type: string
        - name: conservationStatus
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of species
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Species"
    post:
      operationId: createSpecies
      tags: [species]
      summary: Create a species profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpeciesBody"
      responses:
        "201":
          description: Created species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Species"

  /species/{slug}:
    get:
      operationId: getSpecies
      tags: [species]
      summary: Get a species profile by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Species"
        "404":
          description: Not found
    put:
      operationId: updateSpecies
      tags: [species]
      summary: Update a species profile
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpeciesBody"
      responses:
        "200":
          description: Updated species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Species"
    delete:
      operationId: deleteSpecies
      tags: [species]
      summary: Delete a species profile
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /portfolio:
    get:
      operationId: listPortfolioClips
      tags: [portfolio]
      summary: List portfolio clips
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of portfolio clips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioClip"
    post:
      operationId: createPortfolioClip
      tags: [portfolio]
      summary: Create a portfolio clip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioClipBody"
      responses:
        "201":
          description: Created clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioClip"

  /portfolio/{id}:
    delete:
      operationId: deletePortfolioClip
      tags: [portfolio]
      summary: Delete a portfolio clip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactMessageBody"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to the newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsletterSubscriberBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriber"
    get:
      operationId: listNewsletterSubscribers
      tags: [newsletter]
      summary: List newsletter subscribers
      responses:
        "200":
          description: List of subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsletterSubscriber"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required: [id, name, slug]

    CreateTagBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required: [name, slug]

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, createdAt]

    CategoryWithArticles:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
      required: [id, name, slug, createdAt, articles]

    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
      required: [name, slug]

    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        categoryId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        imageAlt:
          type: string
          nullable: true
        readTime:
          type: integer
        wordCount:
          type: integer
          nullable: true
        featured:
          type: boolean
        status:
          type: string
        viewCount:
          type: integer
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        publishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, excerpt, content, category, readTime, featured, status, viewCount, tags, publishedAt, updatedAt, createdAt]

    CreateArticleBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        categoryId:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        imageAlt:
          type: string
          nullable: true
        readTime:
          type: integer
        wordCount:
          type: integer
          nullable: true
        featured:
          type: boolean
        status:
          type: string
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        tagIds:
          type: array
          items:
            type: integer
          nullable: true
      required: [title, slug, excerpt, content, category, readTime, featured, status]

    ArticleStats:
      type: object
      properties:
        total:
          type: integer
        byCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
        featuredCount:
          type: integer
        totalViews:
          type: integer
      required: [total, byCategory, featuredCount, totalViews]

    Species:
      type: object
      properties:
        id:
          type: integer
        commonName:
          type: string
        scientificName:
          type: string
        slug:
          type: string
        kingdom:
          type: string
          nullable: true
        speciesClass:
          type: string
          nullable: true
        orderName:
          type: string
          nullable: true
        family:
          type: string
          nullable: true
        conservationStatus:
          type: string
          nullable: true
        habitat:
          type: string
          nullable: true
        geographicRange:
          type: string
          nullable: true
        diet:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        funFacts:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        iucnUrl:
          type: string
          nullable: true
        articleId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, commonName, scientificName, slug, createdAt]

    CreateSpeciesBody:
      type: object
      properties:
        commonName:
          type: string
        scientificName:
          type: string
        slug:
          type: string
        kingdom:
          type: string
          nullable: true
        speciesClass:
          type: string
          nullable: true
        orderName:
          type: string
          nullable: true
        family:
          type: string
          nullable: true
        conservationStatus:
          type: string
          nullable: true
        habitat:
          type: string
          nullable: true
        geographicRange:
          type: string
          nullable: true
        diet:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        funFacts:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        iucnUrl:
          type: string
          nullable: true
        articleId:
          type: integer
          nullable: true
      required: [commonName, scientificName, slug]

    PortfolioClip:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        publication:
          type: string
        url:
          type: string
        date:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, publication, url, date, category, createdAt]

    CreatePortfolioClipBody:
      type: object
      properties:
        title:
          type: string
        publication:
          type: string
        url:
          type: string
        date:
          type: string
        category:
          type: string
      required: [title, publication, url, date, category]

    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, subject, message, createdAt]

    CreateContactMessageBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
      required: [name, email, subject, message]

    NewsletterSubscriber:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, createdAt]

    CreateNewsletterSubscriberBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
          nullable: true
      required: [email]