# lib/api-zod/src/generated/types/articleStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Writing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleStatsByCategoryItem } from "./articleStatsByCategoryItem";

export interface ArticleStats {
  total: number;
  byCategory: ArticleStatsByCategoryItem[];
  featuredCount: number;
  totalViews: number;
}