# lib/api-zod/src/generated/types/categoryWithArticles.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Writing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";

export interface CategoryWithArticles {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  color?: string | null;
  icon?: string | null;
  createdAt: Date;
  articles: Article[];
}