# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Writing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArticleBody {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  categoryId?: number | null;
  imageUrl?: string | null;
  imageAlt?: string | null;
  readTime: number;
  wordCount?: number | null;
  featured: boolean;
  status: string;
  seoTitle?: string | null;
  seoDescription?: string | null;
  tagIds?: number[] | null;
}