# lib/api-zod/src/generated/types/createCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Writing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryBody {
  name: string;
  slug: string;
  description?: string | null;
  color?: string | null;
  icon?: string | null;
}